Fix wall probs and reorg code
